updated userid to long

+++ Backend/Schichtpilot/Interfaces/IAbsenceService.cs
@@ -5,9 +5,9 @@ namespace Schichtpilot.Interfaces;
 public interface IAbsenceService
 {
     // Employee UCs
-    Task CreateAbsenceRequestAsync(CreateAbsenceDto dto, int userId); // UC-02-01-01
-    Task DeleteOwnAbsenceAsync(int id, int userId); // UC-02-01-02
-    Task<QueryableAbsenceResponse> ViewOwnAbsencesAsync(PaginationDto pagination, AbsenceFilterDto? filter, int userId); // UC-02-01-04
+    Task CreateAbsenceRequestAsync(CreateAbsenceDto dto, long userId); // UC-02-01-01
+    Task DeleteOwnAbsenceAsync(int id, long userId); // UC-02-01-02
+    Task<QueryableAbsenceResponse> ViewOwnAbsencesAsync(PaginationDto pagination, AbsenceFilterDto? filter, long userId); // UC-02-01-04
     
     // Manager UCs  
     Task<QueryableAbsenceResponse> ViewAllAbsencesAsync(PaginationDto pagination, AbsenceFilterDto? filter); // UC-02-01-03
